Add dag.subset() method
This method will be used by deployment module to restrict components
which go into the build/apply steps. If the method is given a list of
initial component names, the new dag will include those components and
all of their ancestors. If the list of components is None, it will
return a copy of the dag object.