Clear component locks before on_build()/on_apply()
The only possible resource contention stemming from multi-threading is
when two links connected to the same destination are executed in
parallel. This means that component's _torque_lock is needed only
until on_build()/on_apply() is about to be executed. We can safely
clear the lock in that moment thus reducing posible lock contention.